Moved localisation, uncommented James' code
Moved localisation to separate class. 
IMPORTANT: still need to decide on mechanism of passing robot information. Created many get methods, may need to be deleted.
