+++ Assignment1/bin/.gitignore
@@ -1,1 +1,2 @@
+/behaviours/
 /net/
